% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/updater.R
\docType{class}
\name{adagrad.updater-class}
\alias{adagrad.updater}
\alias{adagrad.updater-class}
\title{adagrad updater}
\description{
An updater with adaptive step sizes. Adagrad allows different
weights to have different effective learning rates, depending on how
much that parameter has moved so far.
}
\details{
__
}
\section{Fields}{

\describe{
\item{\code{learning.rate}}{the learning rate (set to one in the original paper)}

\item{\code{squared.grad}}{a matrix summing the squared gradients over all previous
updates}

\item{\code{delta}}{the delta matrix (see \code{updater})}
}}

