% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updater.R
\docType{class}
\name{adam.updater-class}
\alias{adam.updater-class}
\alias{adam.updater}
\title{adam updater}
\description{
An updater with adaptive step sizes. Adam allows different 
weights to have different effective learning rates, depending on how 
much that parameter has moved so far and on how much it has moved recently 
in one consistent direction.
}
\section{Fields}{

\describe{
\item{\code{a_0}}{initial step size; default is 0.01}

\item{\code{annealing_rate}}{controls the step size at time \code{t}. Step size is 
\code{a[t] = a_0 / sqrt(1 - annealing_rate + t*annealing_rate)}.
Default is 0.001.}

\item{\code{b1}}{exponential decay rate for first moment estimate; default is 0.9}

\item{\code{b2}}{exponential decay rate for second moment estimate; default is 0.999}

\item{\code{e}}{epsilon (prevents divide-by-zero errors); default is 1E-8}

\item{\code{m}}{first moment estimates; all zero by default at initialization}

\item{\code{v}}{second moment estimates; all zero by default at initialization}

\item{\code{t}}{timestep; zero by default at initialization}

\item{\code{delta}}{the delta matrix (see \code{updater})}
}}


