% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-creation.R
\name{defineLayer}
\alias{defineLayer}
\title{Define a layer's properties}
\usage{
defineLayer(nonlinearity, size, prior, ...)
}
\arguments{
\item{nonlinearity}{a \code{nonlinearity} object describing the activation
function used in this layer}

\item{size}{an integer describing the number of output neurons in this layer.}

\item{prior}{\code{prior} object for constraining the layer's weights.}

\item{...}{additional arguments to be passed to or from other methods 
(currently not used).}
}
\description{
This function produces a description of a \code{layer} in a \code{network}
}
\seealso{
\code{\link{layer}}
}

