% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{importanceSamplingEvaluation}
\alias{importanceSamplingEvaluation}
\title{Evaluate a network object's predictions on newdata against observed y.}
\usage{
importanceSamplingEvaluation(object, newdata, y, loss, batches, batch.size)
}
\arguments{
\item{object}{a \code{network} object}

\item{newdata}{a matrix of predictor variables.}

\item{y}{a matrix of response variables}

\item{loss}{a \code{loss} function (e.g. from a \code{loss} object), assumed
to be a *negative* log-likelihood}

\item{batches}{the number of importance sampling batches to perform}

\item{batch.size}{the number of importance samples per batch}
}
\description{
Estimate the model's log-likelihood on data held out of sample.
Rather than storing a huge number of sample predictions in memory, we can
do this in #' batches of a specified size and just work with the scalar likelihood.
}

