% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\docType{class}
\name{gaussian.prior-class}
\alias{gaussian.prior-class}
\alias{gaussian.prior}
\alias{laplace.prior-class}
\alias{laplace.prior}
\alias{logistic.prior-class}
\alias{logistic.prior}
\alias{lognormal.prior-class}
\alias{lognormal.prior}
\alias{gp.prior-class}
\alias{gp.prior}
\title{\code{gaussian.prior} Gaussian prior with mean and sd}
\description{
\code{gaussian.prior} Gaussian prior with mean and sd

\code{laplace.prior} Laplace "double-exponential" prior with location and scale

\code{logistic.prior} logistic prior distribution with location and scale

\code{lognormal.prior} lognormal prior with meanlog and sdlog parameters

\code{gp.prior} Gaussian process prior
}

