% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updater.R
\docType{class}
\name{rmsprop.updater-class}
\alias{rmsprop.updater-class}
\alias{rmsprop.updater}
\title{rmsprop updater}
\description{
Another updater with adaptive step sizes, like adagrad and adadelta.
}
\details{
https://climin.readthedocs.org/en/latest/rmsprop.html
}
\section{Fields}{

\describe{
\item{\code{learning.rate}}{the learning rate (set to one in the original paper)}

\item{\code{squared.grad}}{a matrix summing the squared gradients over previous
updates (decays according to gamma)}

\item{\code{decay}}{how quickly should squared gradients decay?}

\item{\code{delta}}{the delta matrix (see \code{updater})}
}}


