% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampler.R
\docType{class}
\name{sampler-class}
\alias{gaussian.sampler}
\alias{gaussian.sampler-class}
\alias{sampler}
\alias{sampler-class}
\title{Samplers}
\description{
Samplers produce Monte Carlo samples for \code{\link{network}} objects.
}

