% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{sgd.updater-class}
\alias{sgd.updater}
\alias{sgd.updater-class}
\title{Stochastic gradient descent updater}
\description{
An updater for descending a gradient with momentum
}
\details{
__
}
\section{Fields}{

\describe{
\item{\code{momentum}}{the momentum term}

\item{\code{learning.rate}}{the learning rate}

\item{\code{delta}}{the delta matrix (see \code{updater})}
}}

