% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{updater-class}
\alias{updater}
\alias{updater-class}
\title{Updaters}
\description{
Classes for determining the optimization step to make given a
 gradient
}
\details{
Updaters all have a \code{computeDelta} method, which determines the
 changes in coefficient values to make based on the estimated graidient and
 the current state of the updater.  These changes are stored in a field
 called \code{delta}.

 By default, the \code{mistnet} function uses the same parameters for all
 \code{updaters} in the network, but the user can tune them independently.
}

