% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{zeros}
\alias{zeros}
\title{Produce a matrix of zeros}
\usage{
zeros(n.row, n.col)
}
\arguments{
\item{n.row}{the number of rows desired}

\item{n.col}{the number of columns desired}
}
\description{
Produce a matrix of zeros
}

