% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dropoutMask}
\alias{dropoutMask}
\title{Produce a dropout mask}
\usage{
dropoutMask(nrow, ncol)
}
\arguments{
\item{nrow}{the number of rows desired}

\item{ncol}{the number of columns desired}
}
\description{
Produces a binary matrix, with each element independently sampled as 1 or 0
with probabilty 0.5.
Code is based on the "josh" function from
https://gist.github.com/sckott/3639688
}

