% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loss.R
\docType{data}
\name{loss}
\alias{bernoulliLoss}
\alias{bernoulliRegLoss}
\alias{binomialLoss}
\alias{loss}
\alias{poissonLoss}
\alias{squaredLoss}
\title{Loss functions}
\format{\preformatted{ NULL
}}
\usage{
loss

bernoulliLoss()

bernoulliRegLoss(a, b = a)

poissonLoss()

squaredLoss()

binomialLoss(n)
}
\arguments{
\item{a}{the \code{a} shape parameter in \code{\link{dbeta}}}

\item{b}{the \code{b} shape parameter in \code{\link{dbeta}}}

\item{n}{specifies the number of Bernoulli trials (\code{size} in
\code{\link{dbinom}})}
}
\description{
Loss functions are minimized during model training. \code{loss} objects contain
 a \code{loss} function as well as a \code{grad} function, specifying the gradient.
 Many \code{loss} objects do not require any arguments.

\code{bernoulliLoss}:
cross-entropy for 0-1 data. Equal to
  \code{-(y * log(yhat) + (1 - y) * log(1 - yhat))}

\code{bernoulliRegLoss}: cross-entropy loss, regularized by a
  beta-distributed prior

\code{poissonLoss}: loss based on the Poisson likelihood.
  See \code{\link{dpois}}

\code{squaredLoss}: Squared error, for linear models

\code{binomialLoss}: loss for binomial responses.
}
\keyword{datasets}

