% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{network-class}
\alias{network}
\alias{network-class}
\title{Network}
\description{
A reference class object for a mistnet \code{network} object.
}
\details{
__
}
\section{Fields}{

\describe{
\item{\code{x}}{a numeric matrix of predictor variables.  One row
per example, one column per predictive feature.}

\item{\code{y}}{a \code{matrix} of responses to \code{x}.  One row per example, one
column per response variable.}

\item{\code{layers}}{a \code{list} of \code{layer} objects}

\item{\code{n.layers}}{an integer corresponding to \code{length(layers)}}

\item{\code{dataset.size}}{an integer corresponding to the number of rows in
\code{x} and \code{y}}

\item{\code{n.minibacth}}{an \code{integer} specifying the number of rows to include
in each stochastic estimate of the likelihood gradient.}

\item{\code{minibatch.ids}}{an \code{integer} vector specifying which rows of the
data set to include in the next estimate of the likelihood gradient}

\item{\code{n.importance.samples}}{an \code{integer}}

\item{\code{importance.weights}}{a numeric matrix containing the weights associated
with the most recent round of importance sampling.  (one row per example,
one column per Monte Carlo sample).}

\item{\code{loss}}{the loss function being optimized (not a \code{loss} object!)}

\item{\code{lossGradient}}{the gradient of the loss function being optimized}

\item{\code{sampler}}{the function used to generate Monte Carlo samples for
importance sampling}

\item{\code{completed.iterations}}{a counter that increments after each iteration
of model fitting}
}}

