% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinearity.R
\docType{class}
\name{nonlinearity-class}
\alias{nonlinearity-class}
\alias{nonlinearity}
\alias{linear.nonlinearity-class}
\alias{linear.nonlinearity}
\alias{sigmoid.nonlinearity-class}
\alias{sigmoid.nonlinearity}
\alias{rectify.nonlinearity-class}
\alias{rectify.nonlinearity}
\alias{leaky.rectify.nonlinearity-class}
\alias{leaky.rectify.nonlinearity}
\alias{elu.nonlinearity-class}
\alias{elu.nonlinearity}
\alias{exp.nonlinearity-class}
\alias{exp.nonlinearity}
\alias{mf_mrf.nonlinearity-class}
\alias{mf_mrf.nonlinearity}
\title{nonlinearities}
\description{
--

\code{linear.nonlinearity}:
identity function

\code{sigmoid.nonlinearity}:
logistic function

\code{rectify.nonlinearity}:
\code{max(0, x)}

\code{leaky.rectify.nonlinearity}:
\code{ifelse(x>0, x, x * .01)}

\code{elu.nonlinearity}:
\code{ifelse(x > 0, x, exp(x) - 1)}
}

