% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nonlinearity.R
\docType{class}
\name{nonlinearity-class}
\alias{exp.nonlinearity}
\alias{exp.nonlinearity-class}
\alias{linear.nonlinearity}
\alias{linear.nonlinearity-class}
\alias{mf_mrf.nonlinearity}
\alias{mf_mrf.nonlinearity-class}
\alias{nonlinearity}
\alias{nonlinearity-class}
\alias{rectify.nonlinearity}
\alias{rectify.nonlinearity-class}
\alias{sigmoid.nonlinearity}
\alias{sigmoid.nonlinearity-class}
\title{nonlinearities}
\description{
--

\code{linear.nonlinearity}:
identity function

\code{sigmoid.nonlinearity}:
logistic function

\code{rectify.nonlinearity}:
\code{max(0, x)}
}

