% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updater.R
\docType{class}
\name{sgd.updater-class}
\alias{sgd.updater-class}
\alias{sgd.updater}
\title{Stochastic gradient descent updater}
\description{
An updater for descending a gradient with momentum
}
\details{
__
}
\section{Fields}{

\describe{
\item{\code{momentum}}{the momentum term}

\item{\code{learning.rate}}{the learning rate}

\item{\code{delta}}{the delta matrix (see \code{updater})}
}}


